+++ src/mdsm/dedispersion_manager.h
@@ -6,7 +6,7 @@
 
 SURVEY *processSurveyParameters(QString filepath);
 float* initialiseMDSM(SURVEY* input_survey);
-float* next_chunk(unsigned int data_read, unsigned &samples, long long timestamp = 0, long blockRate = 0);
+float* next_chunk(unsigned int data_read, unsigned &samples, double timestamp = 0, double blockRate = 0);
 int start_processing(unsigned int data_read);
 void tearDownMDSM();
